Refactoring: move all the shared code to a common init function
It is not pretty, but it is more or less the only solution I could find
given how the framework works.
